Rename interface mixins
