+++ StaffMod/src/main/resources/avm_staff.mixins.json
@@ -6,11 +6,11 @@
     "defaultRequire": 1
   },
   "client": [
-    "IMinecraftClientAccessorMixin",
+    "IMinecraftClientMixin",
     "MinecraftClientMixin"
   ],
   "mixins": [
-    "IAbstractFurnaceBlockEntityInvokerMixin",
+    "IAbstractFurnaceBlockEntityMixin",
     "LivingEntityMixin"
   ]
 }
